+++ iOS/AirbnbApp/AirbnbApp/Repository/SearchResultRepoitoriable.swift
@@ -5,7 +5,7 @@
 //  Created by 박진섭 on 2022/06/02.
 //
 
-protocol HouseInfoRepoitoriable {
+protocol SearchResultRepoitoriable {
     var networkManager: NetworkManagable? { get }
     func fetchHouseInfo<T: Codable>(endpoint: Endpointable, onCompleted: @escaping (T?) -> Void)
 }
