+++ iOS/AirbnbApp/AirbnbApp/SupportingFiles/SceneDelegate.swift
@@ -14,7 +14,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        let rootViewController = TabBarController()
+        let rootViewController = MainViewController()
         window?.rootViewController = rootViewController
         window?.makeKeyAndVisible()
     }
