+++ iOS/AirbnbApp/AirbnbApp/ViewController/Search/SearchBarController.swift
@@ -1,0 +1,12 @@
+//
+//  SearchBarController.swift
+//  AirbnbApp
+//
+//  Created by 박진섭 on 2022/06/06.
+//
+
+import UIKit
+
+final class SearchBarController: UIViewController {
+    
+}
